refactor source location about japan module
